% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifd.R
\name{ifd}
\alias{ifd}
\title{Image File Directory}
\usage{
ifd(x, offset = 1)
}
\arguments{
\item{x}{a raw vector.}

\item{offset}{an integer to indicate where the sequence starts.}
}
\value{
a list with the IFD information.
}
\description{
Image File Directory
}
\details{
The output list contains:
\itemize{
\item nb_entries: the number of entries
\item entries: the entries
\item offset_next_ifd: offset to the next ifd.
}
}
\examples{
\dontrun{
ifd(x, offset = 924)
}
}
