% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_frequent.R
\name{most_frequent}
\alias{most_frequent}
\title{Most Frequent}
\usage{
most_frequent(data)
}
\arguments{
\item{data}{a data.frame (output of \code{\link[=scan]{scan()}}).}
}
\value{
a data.frame.
}
\description{
Extract most frequent combination.
}
\details{
The function computes the most frequent combination over:
camera, orientation, exposure_time, f_number, iso_speed, lens_model, focal_length
}
\examples{
\dontrun{
most_frequent(scan("."))
}
}
