% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_camera.R
\name{p_camera}
\alias{p_camera}
\title{Camera Plot}
\usage{
p_camera(data, fg = NA, bg = "grey", theme = p_theme())
}
\arguments{
\item{data}{a data.frame (see details).}

\item{fg}{a foreground color.}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
Camera Plot
}
\details{
\code{data} expects a data.frame with the following columns:
\itemize{
\item camera: character, the camera name
\item n: numeric, the number of images for the camera
}
}
\examples{
p_camera(data.frame(camera = "Canon EOS 70D", n = 10))
}
\seealso{
\code{\link[=p_theme]{p_theme()}}
}
