% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_focal.R
\name{p_focal}
\alias{p_focal}
\title{Focal Length Plot}
\usage{
p_focal(data, fg = "grey", bg = "grey", theme = p_theme())
}
\arguments{
\item{data}{a data.frame (see details).}

\item{fg}{a foreground color.}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
Focal Length Plot
}
\details{
\code{data} expects the following columns:
\itemize{
\item lens_model: character, the name of the lens
\item focal_length: numeric, the focal length
\item n: the number of images.
}
}
\examples{
\dontrun{
p_focal(data = metadata |>
dplyr::group_by(lens_model, focal_length) |>
dplyr::summarise(n = dplyr::n()))
}
}
