% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_iso.R
\name{p_iso}
\alias{p_iso}
\title{ISO Plot}
\usage{
p_iso(data, bg = "grey", theme = p_theme())
}
\arguments{
\item{data}{a data.frame (see details).}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
ISO Plot
}
\details{
\code{data} expects the following columns:
\itemize{
\item lens_model: character, the name of the lens
\item iso_speed: numeric, the ISO speed
\item n: the number of images.
}
}
\examples{
\dontrun{
p_iso(data.frame(lens_model = "foo", iso_speed = 100, n = 3))
}
}
