% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_lens.R
\name{p_lens}
\alias{p_lens}
\title{Lens Plot}
\usage{
p_lens(data, fg = "#000", bg = "grey", theme = p_theme())
}
\arguments{
\item{data}{a data.frame (see details).}

\item{fg}{a foreground color.}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
Raw images per lens type.
}
\details{
\code{data} expects the following columns:
\itemize{
\item lens_model: a character string, the name of the lens
\item n: numeric, the number of images for this lens
}
}
\seealso{
\code{\link[=p_theme]{p_theme()}}
}
