% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_orientation.R
\name{p_orientation}
\alias{p_orientation}
\title{Orientation Plot}
\usage{
p_orientation(nl, np, fg = "black", bg = "grey", theme = p_theme())
}
\arguments{
\item{nl}{a numeric, the number of landscape.}

\item{np}{a numeric, the number of portraits.}

\item{fg}{a foreground color.}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
Distribution over landscape and portrait.
}
\examples{
p_orientation(nl = 42, np = 112)
}
\seealso{
\code{\link[=p_theme]{p_theme()}}
}
