% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_overview.R
\name{p_overview}
\alias{p_overview}
\title{Overview Plot}
\usage{
p_overview(metadata)
}
\arguments{
\item{metadata}{a data.frame (see details).}
}
\value{
a ggplot object.
}
\description{
Overview Plot
}
\details{
\code{metadata} expects an output from the \code{\link[=scan]{scan()}} function.
}
\examples{
\dontrun{
viz(scan(path = "."))
}
}
\seealso{
\code{\link[=scan]{scan()}}
}
