% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_shutter_speed.R
\name{p_shutter_speed}
\alias{p_shutter_speed}
\title{Shutter Speed Plot}
\usage{
p_shutter_speed(data, fg = "grey", bg = "grey", theme = p_theme())
}
\arguments{
\item{data}{a data.frame (see details).}

\item{fg}{a foreground color.}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
Shutter Speed Plot
}
\details{
\code{data} expects the following columns:
\itemize{
\item lens_model: character, the name of the lens
\item exposure_time: character, the aperture value
\item n: the number of images.
}
}
\examples{
\dontrun{
p_shutter_speed(data)
}
}
