% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_theme.R
\name{p_theme}
\alias{p_theme}
\title{Default Theme}
\usage{
p_theme()
}
\value{
a 'ggplot' theme.
}
\description{
Default Theme
}
\details{
The function is used as a default theme for the plots, which
can be replaced by a custom one.
}
\examples{
p_theme()
}
