% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_bytes.R
\name{raw_bytes}
\alias{raw_bytes}
\title{Sequence of Bytes}
\usage{
raw_bytes(x, offset = 0, n = 1)
}
\arguments{
\item{x}{a vector of raw bytes.}

\item{offset}{an integer, to indicate the lower limit of the sequence.}

\item{n}{an integer, to indicate the length of the sequence.}
}
\value{
a vector of raw bytes.
}
\description{
Sequence of Bytes
}
\details{
Note that offset = 0 (default) is the index of the first element of the list.
}
\examples{
\dontrun{
# get a sequence of 4 bytes at offset 294
raw_bytes(x, offset = 294, n = 4)
}
}
