% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz}
\alias{viz}
\title{Metadata Visualization}
\usage{
viz(metadata)
}
\arguments{
\item{metadata}{a data.frame of the metadata (output of scan function)}
}
\value{
a ggplot object
}
\description{
Metadata Visualization
}
\examples{
\dontrun{
viz(scan(path = "."))
}
}
