% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{CR2 Header}
\usage{
header(x)
}
\arguments{
\item{x}{a raw vector}
}
\value{
a list of values (TIFF & Canon headers)
}
\description{
CR2 Header
}
\examples{
\dontrun{
header(x)
}
}
