% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifd_entry.R
\name{ifd_entry}
\alias{ifd_entry}
\title{IFD Entry Value or Offset}
\usage{
ifd_entry(ifd, id, raw_vector = NULL, verbose = FALSE)
}
\arguments{
\item{ifd}{an IFD list as return from the ifd function}

\item{id}{the id of the entry}

\item{raw_vector}{the raw vector to extract value when an offset is involved}

\item{verbose}{whether to turn ON/OFF console log}
}
\value{
a numeric value or a vector with the offset
}
\description{
IFD Entry Value or Offset
}
\examples{
\dontrun{
ifd_entry(ifd, id = "0100")
}
}
