% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_bytes.R
\name{order_bytes}
\alias{order_bytes}
\title{Order Bytes}
\usage{
order_bytes(x, endian = "II")
}
\arguments{
\item{x}{a vector or data.frame of bytes}

\item{endian}{the endianess}
}
\value{
the corresponding hexadecimal vector or data.frame
}
\description{
Order Bytes
}
\details{
When x is a data.frame, value will be a data.frame, each row being
the concatenation of the ordered columns
}
\examples{
\dontrun{
order_bytes(x, endian = "II")
}
}
