% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_aperture.R
\name{p_aperture}
\alias{p_aperture}
\title{Aperture Plot}
\usage{
p_aperture(data, bg = "grey", theme = p_theme())
}
\arguments{
\item{data}{a data.frame (see details).}

\item{bg}{a background color.}

\item{theme}{an optional theme function.}
}
\value{
a 'ggplot' object.
}
\description{
Aperture Plot
}
\details{
\code{data} expects the following columns:
\itemize{
\item lens_model: character, the name of the lens
\item f_number: character, the aperture value
\item n: the number of images.
}
}
\examples{
\dontrun{
p_aperture(data)
}
}
