% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cr2.R
\name{read_cr2}
\alias{read_cr2}
\title{Read CR2 File}
\usage{
read_cr2(file, image = FALSE, mapping_exif = NULL, mapping_canon = NULL)
}
\arguments{
\item{file}{the file to read (including path)}

\item{image}{a logical whether the images should be loaded or not (default FALSE)}

\item{mapping_exif}{a data.frame of the EXIF tag mapping}

\item{mapping_canon}{a data.frame of the Canon tag mapping}
}
\value{
a list of metadata
}
\description{
Read CR2 File
}
\examples{
\dontrun{
read_cr2(file = "C:/download/example.CR2")
}
}
