% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan}
\alias{scan}
\title{Scan Folder}
\usage{
scan(path)
}
\arguments{
\item{path}{the path of the folder to scan}
}
\value{
a data.frame of the metadata
}
\description{
Scan Folder
}
\examples{
scan(path = ".")
}
