% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_id}
\alias{tag_id}
\title{Tag Id}
\usage{
tag_id(mapping, x)
}
\arguments{
\item{mapping}{a data.frame for the mapping.}

\item{x}{a character string, the name of the key.}
}
\value{
a character string.
}
\description{
Tag Id
}
\examples{
\dontrun{
tag_id(mapping_exif, "ImageWidth")
}
}
